+++ ForecastRestWithLibs/app/src/main/java/com/android2ee/formation/restservice/forecastyahoo/withlibs/transverse/model/clientside/newest/WeatherDay.java
@@ -111,9 +111,9 @@ public class WeatherDay extends SugarRecord {
             clouds = new Clouds(weatherServerSide.getClouds());
             rain = new Rain(weatherServerSide.getRain());
             snow = new Snow(weatherServerSide.getSnow());
-            timeStampUTC=weatherServerSide.getTimeStampUTC();
+            timeStampUTC= (int) weatherServerSide.getTimeStampUTC();
             ephemeris=new Ephemeris(weatherServerSide.getSys());
-            cityId=weatherServerSide.getCityId();
+            cityId= (int) weatherServerSide.getCityId();
 //            weatherMetaData=new WeatherMetaData(weatherServerSide.getMain().
         }
     }
